pft: Enable multi-map support
The code to create multiple memory regions (one per thread) is
available but commented out. The feature is useful to evaluate the
benefits of per-VMA locks[0] patchset which breaks the mmap_lock
covering the entire address space into locks per region of memory.

As the code was able to create multiple regions (verified via
/proc/$$/maps), enable it for wider usage.

Signed-off-by: Punit Agrawal <punit.agrawal@bytedance.com>